feat(frontend): expose the facts the mobile overview cards read
Additive package changes only: agentConfigSummary names the integration behind each gateway-connection tool, NextTriggersSection's row derivation moves into a useUpcomingTriggers hook a host with its own card chrome can call, and composioLogo is exported for any integration key outside PROVIDERS.
